+++ src/rust/src/series.rs
@@ -577,7 +577,7 @@ impl RPolarsSeries {
         Ok(RPolarsSeries(s))
     }
 
-    pub fn into_frame(&self) -> RPolarsDataFrame {
+    pub unsafe fn into_frame(&self) -> RPolarsDataFrame {
         RPolarsDataFrame(pl::DataFrame::new_no_checks(vec![self.0.clone()]))
     }
 }